refactor: remove combobox demo module
