+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow-integration-tests/src/main/java/com/vaadin/flow/component/combobox/test/entity/Project.java
@@ -1,4 +1,4 @@
-package com.vaadin.flow.component.combobox.demo.entity;
+package com.vaadin.flow.component.combobox.test.entity;
 
 public class Project {
